Adds move content functionality to clone content
Enables users to move content by cloning and then deleting the original.
Adds a checkbox to the clone content form to allow users to choose whether to delete the original content after cloning.
Updates the clone content function to delete the original content if the checkbox is checked.
Also handles exceptions that do not have a status.

Fixes ticket_1651089252